refactor: extract rehydrateChildren to a separate file
